--- a/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoCommon.cs
+++ b/src/System.Text.Json/src/System/Text/Json/Serialization/JsonPropertyInfoCommon.cs
@@ -2,6 +2,8 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Collections;
+using System.Collections.Generic;
 using System.Diagnostics;
 using System.Reflection;
 using System.Text.Json.Serialization.Converters;
@@ -82,6 +84,11 @@ namespace System.Text.Json.Serialization
             {
                 Set((TClass)obj, (TDeclaredProperty)value);
             }
+        }
+
+        internal override IList CreateConverterList()
+        {
+            return new List<TDeclaredProperty>();
         }
     }
 }