… src/System.Text.Json/tests/Serialization/TestClasses.SimpleTestClass.cs
… src/System.Text.Json/tests/Serialization/TestClasses.SimpleTestClassWithObject.cs
… src/System.Text.Json/tests/Serialization/TestClasses.cs