Avoids boxing when using primitive arrays in an object. CPU improvment is almost twice as fast due to ~8x fewer allocations caused by boxing.

There are two fixes:
- The "temporary list" used to hold values until it can be converted to an `T[]` array went from `List<object>` to `List<T>` - this avoids boxing by the `ToArray` function.
- When adding elements to the list it doesn't box the primitive `T` value to `object` any longer.

Also tests were refactored to place the large test classes to each have their own file, and expand on the nullable tests since they were missing nullable arrays. There was an issue found due to this (with nullable arrays) and that is also fixed here.

before (`byte[10000]`):
```
|            Method |       Mean |    Error |   StdDev |     Median |         Min |        Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------ |-----------:|---------:|---------:|-----------:|------------:|-----------:|------------:|------------:|------------:|--------------------:|
|          JSON.NET |   799.0 us | 15.33 us | 14.34 us |   798.9 us |   776.57 us |   827.6 us |     66.4557 |      3.1646 |           - |            285720 B |
|  System.Text.Json | 1,345.8 us | 24.13 us | 22.57 us | 1,340.5 us | 1,312.29 us | 1,394.5 us |     76.9231 |     38.4615 |     38.4615 |            532512 B |
```

after:
```
|            Method |      Mean |     Error |     StdDev |    Median |       Min |         Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------ |----------:|----------:|-----------:|----------:|----------:|------------:|------------:|------------:|------------:|--------------------:|
|          JSON.NET | 805.39 us | 39.037 us |  43.390 us | 793.70 us | 753.08 us |   892.85 us |     66.0377 |      3.1447 |           - |            285720 B |
|  System.Text.Json | 759.17 us | 76.111 us |  84.597 us | 717.42 us | 662.12 us |   935.97 us |     13.7363 |           - |           - |             63192 B |
```