@ahsonkhan 2019-04-14
> There was an issue found due to this (with nullable arrays) and that is also fixed here.